+++ client/src/app/domain/models/users/user.constants.ts
@@ -11,20 +11,3 @@ export const userHeadersAndVerboseNames: { [key in keyof User]?: string } = {
     username: `Username`,
     gender: `Gender`
 };
-
-export const accountHeadersAndVerboseNames: { [key in keyof User]?: string } = {
-    ...userHeadersAndVerboseNames,
-    default_number: `Participant number`,
-    default_structure_level: `Structure level`,
-    default_vote_weight: `Vote weight`
-};
-
-export const participantHeadersAndVerboseNames: { [key in keyof User]?: string } = {
-    ...userHeadersAndVerboseNames,
-    number: `Participant number`,
-    structure_level: `Structure level`,
-    vote_weight: `Vote weight`,
-    comment: `Comment`,
-    is_present_in_meeting_ids: `Is present`,
-    group_ids: `Groups`
-};
