+++ client/src/app/site/pages/meetings/pages/participants/pages/participant-import/components/participant-import-list/participant-import-list.component.ts
@@ -2,10 +2,10 @@ import { Component } from '@angular/core';
 import { FormBuilder } from '@angular/forms';
 import { TranslateService } from '@ngx-translate/core';
 import { User } from 'src/app/domain/models/users/user';
-import { participantHeadersAndVerboseNames } from 'src/app/domain/models/users/user.constants';
 import { BaseUserImportListComponent } from 'src/app/site/base/base-user-import-list.component';
 import { MeetingComponentServiceCollectorService } from 'src/app/site/pages/meetings/services/meeting-component-service-collector.service';
 
+import { participantHeadersAndVerboseNames } from '../../definitions';
 import { ParticipantImportService } from '../../services';
 
 @Component({
