+++ client/src/app/site/pages/organization/pages/accounts/pages/account-import/components/account-import-list/account-import-list.component.ts
@@ -2,11 +2,11 @@ import { Component, OnInit } from '@angular/core';
 import { FormBuilder } from '@angular/forms';
 import { TranslateService } from '@ngx-translate/core';
 import { User } from 'src/app/domain/models/users/user';
-import { accountHeadersAndVerboseNames } from 'src/app/domain/models/users/user.constants';
 import { BaseUserImportListComponent } from 'src/app/site/base/base-user-import-list.component';
 import { ComponentServiceCollectorService } from 'src/app/site/services/component-service-collector.service';
 
 import { AccountControllerService } from '../../../../services/common/account-controller.service';
+import { accountHeadersAndVerboseNames } from '../../definitions';
 import { AccountImportService } from '../../services/account-import.service/account-import.service';
 
 @Component({
