+++ client/src/app/site/pages/organization/pages/accounts/pages/account-import/definitions/index.ts
@@ -1,0 +1,9 @@
+import { User } from 'src/app/domain/models/users/user';
+import { userHeadersAndVerboseNames } from 'src/app/domain/models/users/user.constants';
+
+export const accountHeadersAndVerboseNames: { [key in keyof User]?: string } = {
+    ...userHeadersAndVerboseNames,
+    default_number: `Participant number`,
+    default_structure_level: `Structure level`,
+    default_vote_weight: `Vote weight`
+};
