+++ client/src/app/site/pages/organization/pages/accounts/pages/account-import/services/account-import.service/account-import.service.ts
@@ -1,12 +1,12 @@
 import { Injectable } from '@angular/core';
 import { User } from 'src/app/domain/models/users/user';
-import { accountHeadersAndVerboseNames } from 'src/app/domain/models/users/user.constants';
 import { ImportConfig } from 'src/app/infrastructure/utils/import/import-utils';
 import { BaseUserImportService } from 'src/app/site/base/base-user-import.service';
 import { ImportServiceCollectorService } from 'src/app/site/services/import-service-collector.service';
 import { UserControllerService } from 'src/app/site/services/user-controller.service';
 
 import { AccountExportService } from '../../../../services/account-export.service';
+import { accountHeadersAndVerboseNames } from '../../definitions';
 import { AccountImportServiceModule } from '../account-import-service.module';
 
 @Injectable({
